Interface: b2PolygonShape: getNormals_b2Vec2, getNormals_tuple added b2Vec2.dot(v2) added.
Miscellaneous:
fluiDemo: Port of Blaze's fluid physics. Too slow in Python to be used in a game. More of a proof of concept. Mostly works, some kinks still there though.

UFOCatcher: Sprites working with Rabbyt, looks pretty but the grip is still not great.

Both require a copy of the data/pgu directories from trunk/testbed to work.