Attempting to do the whole unit test thing this time around. I'm sure with the 5 minutes I read the manual it's not the most optimal way. So if you're actually reading these commit messages, let me know of the err of my ways.
On another note, getting rid of the getAsType() nonsense, b2World.CreateJoint() returns a
properly casted joint on its own. Will do it for shapes next time around.

I'll leave the new version, downcast() in there, though it's not really necessary.

New swig is mentioning b2Fixture* has no destructor (a memory leak, it claims), but I'll
have to look into the validity of it, and disable that warning if it's false.